Resolve implicit 64->32 bit castings in Adium.app

+++ Frameworks/AIUtilities Framework/Source/AIArrayAdditions.h
@@ -15,6 +15,6 @@
 
 @interface NSMutableArray (ESArrayAdditions)
 - (void)addObjectsFromArrayIgnoringDuplicates:(NSArray *)inArray;
-- (void)moveObject:(id)object toIndex:(unsigned)newIndex;
-- (void)setObject:(id)object atIndex:(unsigned)idx;
+- (void)moveObject:(id)object toIndex:(NSUInteger)newIndex;
+- (void)setObject:(id)object atIndex:(NSUInteger)idx;
 @end
